% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInformation.R
\name{imageInformation}
\alias{imageInformation}
\alias{imageinformation}
\title{imageInformation}
\usage{
imageInformation(input_dir = NULL, input_file = NULL)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{input_file}{A character (image file)}
}
\description{
Main function for getting image information
}
\details{
Input should be tif-format.
}
\author{
Kai Budde
}
